requirements.txt
# Bringetto Trading Agent - Python Dependencies

# Alpaca Trading API
alpaca-py>=0.21.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Data Processing
pandas>=2.2.0
numpy>=1.26.3
ta>=0.11.0

# Machine Learning
scikit-learn>=1.4.0

# Scheduling & Background Tasks
apscheduler>=3.10.4
redis>=5.0.1

# Data Validation
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Logging
loguru>=0.7.2

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.1

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.2
pytz>=2024.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
ruff>=0.1.14
mypy>=1.8.0

# Types
types-python-dateutil>=2.8.19
types-pytz>=2024.1.0
types-redis>=4.6.0
